Updated Some Things
src/lxmusic.c

Removed the handler's for cut/copy/paste, as if it's not implemented,
then it's not a needed "feature".

Changed some text in the About window:

Rearranged some text from "A Simple GUI XMMS2 client" to "A Simple XMMS2
GUI Client", as it sounds better if you were to say it out loud.

Added some more text to the License popup, mainly telling the user where
they can go to read the GNU GPL 2.0 License, and changing the URL for
the LxDE page from "http://lxde.org" to "https://www.lxde.org".

Changed the text "Show Main Window" to "Show LXMusic", not really
something that matters, more of an aesthetic change

data/lxmusic.ui.glade

Removed the Cut/Paste buttons, for the same reason in the 'lxmusic.c'
file, if it's not implemented, then it's not needed, as it cannot be
utilised, additionally removed the 'Tools' tab, as in my testing it
showed no sub-menu options.

Moved the 'About' menu option to the 'File' menu

-------------------------------------------
  Contact: transsentientbeing@outlook.com
-------------------------------------------